config/settings.py
import os
from dotenv import load_dotenv

load_dotenv()

class Settings:
    GROQ_API_KEY = os.getenv("GROQ_API_KEY")
    DEBUG = os.getenv("DEBUG", "False") == "True"

settings = Settings()
