requirements.txt
langchain
langchain-community
langchain-groq
streamlit
python-dotenv
boto3